Support for default options.
This works even if the file doesn't work, and across multiple
invocations (even if cached).

The cache step has been moved a step back, so that the parsed YAML is
cached, but the resolved c-c-config is not. This allows different
defaults to be passed in for each call.

In addition, caching now ensures a config file that doesn't exist won't
be checked for again, until the cache is cleared.